requirements.txt
langchain==0.1.20
streamlit==1.34.0
boto3
pdfplumber
python-dotenv
